catalog_engine/producer-consumer/requirements.txt
confluent-kafka==1.9.0
psycopg2-binary==2.9.3